Voice_Assistant/requirements.txt
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
fuzzywuzzy==0.18.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
outcome==1.3.0.post0
packaging==25.0
psutil==7.0.0
pvporcupine==3.0.5
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.1.1
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
screen_brightness_control==0.24.2
selenium==4.34.2
setuptools==80.9.0
silero==0.4.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
srt==3.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.53.2
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
vosk==0.3.45
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
